feat: seperated physical & encrypted wallet recovery flow
