+++ lib/recover_wallet/ui/screens/physical/recover_physical_wallet_success_screen.dart
@@ -8,8 +8,8 @@ import 'package:gap/gap.dart';
 import 'package:gif/gif.dart';
 import 'package:go_router/go_router.dart';
 
-class RecoverWalletSuccessScreen extends StatelessWidget {
-  const RecoverWalletSuccessScreen({super.key});
+class RecoverPhysicalWalletSuccessScreen extends StatelessWidget {
+  const RecoverPhysicalWalletSuccessScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
